pyproject.toml
[project]
name = "marketos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.0",
    "celery[redis]>=5.0",
    "firecrawl-py",
    "groq",
    "google-cloud-storage",
    "google-cloud-aiplatform",
    "pydantic",
    "python-dotenv",
    "requests",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]

